🔧 Fix remaining Qodana annotations across 21 files
- Remove redundant nullable return types (Language?, Icon?)
- Remove redundant .toShort() conversion and \} regex escape
- Delete unused symbols (TOKEN_SEPARATOR, EXPR_PATTERN, unused params)
- Add @Suppress for unused public API constants (TOOL_WINDOW_ID, INSTANCE)
- Add @Suppress for unstable API usage (Alarm, CodeVisionPlaceholderCollector)
- Add @Suppress for string capitalization ("PPX Annotations")
- Fix var→val for never-reassigned variable
- Simplify call chains (ifEmpty, joinToString with transform)
- Remove unused local variables in handlers

+++ src/main/kotlin/com/rescript/plugin/hierarchy/call/RescriptCallAnalyzer.kt
@@ -66,7 +66,7 @@ object RescriptCallAnalyzer {
             },
             scope,
             name,
-            UsageSearchContext.IN_CODE.toShort(),
+            UsageSearchContext.IN_CODE,
             true, // case sensitive
         )
